backend/requirements.txt
aiohttp==3.8.4
python-slugify==8.0.1
python-socketio==5.8.0
python-magic==0.4.27